app/requirements.txt
flask
pymysql
rapidfuzz
python-dotenv
